Load scRNA-seq / RNA-seq Data

This function processes multiple scRNA-seq / RNA-seq files, filters based on annotation types, and concatenates data as specified.

load_rnaseq(
  path,
  concatenation_type = "sum",
  annotation_sides = c("three_prime_UTR", "exon", "five_prime_UTR")
)

Arguments

path

Path to the directory containing scRNA-seq / RNA-seq matrices.

concatenation_type

Method to concatenate data: 'sum', 'avr', or 'top_diff'.

annotation_sides

A vector of annotation sides to include (e.g., 'three_prime_UTR', 'exon', 'five_prime_UTR').

Value

A list of data frames, each corresponding to an annotation type.

Examples

load_sc_rnaseq("data/matrices/", concatenation_type = 'sum')
#> Error in load_sc_rnaseq("data/matrices/", concatenation_type = "sum"): could not find function "load_sc_rnaseq"