Normalize gene expression data

Applies log2 normalization to gene expression data based on gene counts and a scaling factor.

normalize_data(data, type = "counts", factor = 1e+06)

Arguments

data

A matrix or data frame of raw gene expression data.

type

A type of normalization based on number of positive expressed genes or total counts per cell / set. Default is 'counts'

factor

A numeric scaling factor for normalization. Default is 1,000,000.

Value

A normalized dataset with log2-transformed values.

Examples

normalized_data <- normalize_data(data)
#> Error in normalize_data(data): could not find function "normalize_data"